Testing default date open
Initial value for forms is set in the form which is set at the time of server restart, so after 2 days "today" becomes 2 days ago. Testing setting the initial value in the views.py instead. This should take the real datetime.datetime.now() as views are generated at time of request, not server start. Only changing opening items for now incase it doesn't work, if it does other forms will be changed later.