Split long items in PDF reports
If an item is longer than 40 charaters it gets split over two lines. 
Also had to change how page numbers are counted due to this (previously 
was just lines//29), now it builds a fake report (with obvious headers 
just incase it gets shown somehow) that it uses to count the total 
number of pages then makes the real report with that number.