Fix for solutions that track volumes but have components that don't
Previously if a solution had it's volume tracked it was assumed that all 
reagents track volumes. This caused a problem as it sometimes tried to 
subtract an amount used from an item that didn't have a value for 
ammount used. Now if a recipe has items that don't track volumes those 
items can't have volumes used