Added teams
Teams are who is responsible for an item. Default teams are "DNA" and 
"CYTO".

Added teams to priming script.

Teams behave similar to suppliers, in that each reagent has a default 
one but it can be changed at the time of booking in.

Teams can be (de)activated and default team can be changed.

Added teams filtering to search and inventory reports