Changes to Solution Validation
The "****NOT VALIDATED****" Message on the summary page only shows up 
for solutions that have components that are not validated.

Having a ~ in the name does not automatically validate an item if it is 
a solution.

Trying to validate a solution by manually going to /stock/valitem/<ITEM> 
redirects you to the items summary page if it is a solution