Can validate finished items.
So long as an item has been opened it can be validated, even if the item 
is finished.

Also changed it so you can un-validate a finished item (in editinv)

Also fixed undoing a discarded item, previously it's action was 
"unopen", which was incorrect as it did not return it to an open state. 
Added a new "undiscard" section to undo function as the open count 
(7f1f85822c31be869be742c1d2b0052ab1210785) breaks if using re-open for 
something that doesn't get opened.