If an email fails to send it will be saved in "Emails" table to send later
If an Email fails to send the subject, send to and text get stored in the "Emails" table. APScheduler then, every 30 minutes, checks if there are unsent emails and if there are, sends them. If Scheduler succeeds at sending email it updates sent=True, otherwise stays at False and gets retried next 30 mins