Fix for items that have two "/" in their names
On the search page (also accessed by clicked an item from main page), if 
an item has a "/" in the name that broke the page as the URL would not 
be split correctly. Was already a fix in place for items with one "/" in 
their names, however if two "/" were used in an item it would
returned no results. This fix should would for any number of "/" in an 
items name.