Beginnings of kit insert information
Added a page to view kit inserts. From here Admins can also add new kit 
insert versions.

The latest version of a kit insert shows in an items summary page

Added a search to look for items recieved between certain dates that 
have kit inserts at various stages of checking.

If the location of a kit insert is a valid web address it will be a 
clickable hyperlink